Restructure packaging of library jar; bundling via shadowJar is no longer necessary.
This reverts commit 032d73c6a17ffb6a6094fa7a1665f23bd277f7cd.
