+++ sri-efactura-model/build.gradle.kts
@@ -3,7 +3,7 @@
 val jaxXjc: Configuration by configurations.creating
 
 plugins {
-    id("sri-efactura-utils.java-library-conventions")
+    id("sri-efactura-core.java-library-conventions")
     idea
 }
