+++ sri-soap-cli/build.gradle.kts
@@ -1,12 +1,15 @@
 
 plugins {
-    id("sri-efactura-utils.kotlin-application-conventions")
+    id("sri-efactura-core.kotlin-application-conventions")
 }
 
 dependencies {
     implementation(project(":sri-client"))
-    implementation("com.fasterxml.jackson.core:jackson-databind:2.13.3")
+    implementation("com.fasterxml.jackson.core:jackson-databind:2.14.0")
     implementation("com.offbytwo:docopt:0.6.0.20150202")
+}
+repositories {
+    mavenCentral()
 }
 
 application {
